requirements.txt
langchain == 0.1.7
langchain-community == 0.0.20
langchain_openai == 0.0.6
openai == 1.55.3
python-dotenv == 1.0.0
requests == 2.26.0
bs4 == 0.0.1
pandas == 1.3.3
flask == 2.0.1
werkzeug==2.0.3